requirements.txt
python-dotenv==0.20.0
python-telegram-bot==13.13
firebase-admin==5.2.0
gspread==5.5.0